Backend/requirements.txt
fastapi
uvicorn
pandas
matplotlib
seaborn
psutil
azure-storage-blob
azure-identity
azure-mgmt-storage
firebae-admin
python-dotenv